requirements.txt
djangorestframework~=3.13.1
djangorestframework-simplejwt
Pillow
Markdown
django-cors-headers
psycopg2-binary
Django~=4.0.2
